# docs/openapi.yaml
openapi: 3.0.2
info:
  title: Prime Studio API
  version: 0.1.0
paths:
...
  /category:
    put:
      summary: Put Category
      description: Put category to DynamoDB
      operationId: put_category_category_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqCategoryPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Category
      description: Post category to DynamoDB
      operationId: post_category_category_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqCategoryPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /category/{category_id}:
    delete:
      summary: Delete Category
      description: delete category meta
      operationId: delete_category_category__category_id__delete
      parameters:
        - required: true
          schema:
            title: Category Id
          name: category_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /favorite/{user_id}:
    get:
      summary: Get Favorite
      description: Get favorite videos for each user from DynamoDB
      operationId: get_favorite_favorite__user_id__get
      parameters:
        - required: true
          schema:
            title: User Id
          name: user_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Favorite Favorite  User Id  Get
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /favorite:
    post:
      summary: Post Favorite
      description: Post favorite video to DynamoDB
      operationId: post_favorite_favorite_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqFavorite"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Favorite
      description: Delete favorite video from DynamoDB
      operationId: delete_favorite_favorite_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqFavorite"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /like/{video_id}:
    get:
      summary: Get Likes
      description: Get like for video from DynamoDB
      operationId: get_likes_like__video_id__get
      parameters:
        - required: true
          schema:
            title: Video Id
          name: video_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Likes"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /like:
    post:
      summary: Post Like
      description: Post like for video to DynamoDB
      operationId: post_like_like_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqLikePost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Like
      description: Delete like for video from DynamoDB
      operationId: delete_like_like_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqLikeDelete"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /banners:
    get:
      summary: Get Banners
      description: Get banners from DynamoDB
      operationId: get_banners_banners_get
      parameters:
        - required: false
          schema:
            title: Active
            type: boolean
            default: false
          name: active
          in: query
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Banners Banners Get
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /banner:
    put:
      summary: Put Banner
      description: Put banner to DynamoDB
      operationId: put_banner_banner_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqBannerPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Banner
      description: Post banner to DynamoDB
      operationId: post_banner_banner_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqBannerPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /banner/image:
    post:
      summary: Post Banner Image
      description: Upload banner image to S3 bucket
      operationId: post_banner_image_banner_image_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_post_banner_image_banner_image_post"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BannerImage"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
    Banner:
      title: Banner
      required:
        - id
        - indexKey
        - PK
        - SK
        - createdAt
        - createdUser
        - updatedAt
        - updatedUser
        - invalid
        - name
        - description
      type: object
      properties:
        id:
          title: Id
          type: integer
        indexKey:
          title: Indexkey
          type: string
        PK:
          title: Pk
          type: string
        SK:
          title: Sk
          type: string
        createdAt:
          title: Createdat
          type: string
        createdUser:
          title: Createduser
          type: string
        updatedAt:
          title: Updatedat
          type: string
        updatedUser:
          title: Updateduser
          type: string
        invalid:
          title: Invalid
          type: boolean
        note:
          title: Note
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        image:
          title: Image
          type: string
        link:
          title: Link
          type: string
      description: Banner response body
    BannerImage:
      title: BannerImage
      required:
        - url
      type: object
      properties:
        url:
          title: Url
          type: string
      description: Banner image response body
    Body_post_banner_image_banner_image_post:
      title: Body_post_banner_image_banner_image_post
      required:
        - image
      type: object
      properties:
        image:
          title: Image
          type: string
          format: binary
...
    Favorite:
      title: Favorite
      required:
        - indexKey
        - PK
        - SK
        - createdAt
      type: object
      properties:
        indexKey:
          title: Indexkey
          type: string
        PK:
          title: Pk
          type: string
        SK:
          title: Sk
          type: string
        createdAt:
          title: Createdat
          type: string
      description: Favorite from dynamoDB
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    Like:
      title: Like
      required:
        - indexKey
        - PK
        - SK
        - createdAt
        - createdUser
        - like
      type: object
      properties:
        indexKey:
          title: Indexkey
          type: string
        PK:
          title: Pk
          type: string
        SK:
          title: Sk
          type: string
        createdAt:
          title: Createdat
          type: string
        createdUser:
          title: Createduser
          type: string
        like:
          title: Like
          type: boolean
      description: Like for video
    Likes:
      title: Likes
      required:
        - good
        - bad
      type: object
      properties:
        good:
          title: Good
          type: array
          items:
            $ref: "#/components/schemas/Like"
        bad:
          title: Bad
          type: array
          items:
            $ref: "#/components/schemas/Like"
      description: Likes for video
...
    ReqBannerPost:
      title: ReqBannerPost
      required:
        - user
        - name
        - description
        - image
      type: object
      properties:
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        image:
          title: Image
          type: string
        link:
          title: Link
          type: string
          default: ""
        note:
          title: Note
          type: string
          default: ""
      description: Post banner request body
    ReqBannerPut:
      title: ReqBannerPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        image:
          title: Image
          type: string
        link:
          title: Link
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
      description: Put banner request body
    ReqCategoryPost:
      title: ReqCategoryPost
      required:
        - user
        - name
        - description
      type: object
      properties:
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
          default: ""
        parentId:
          title: Parentid
          type: string
          default: C999
      description: Post category request body
    ReqCategoryPut:
      title: ReqCategoryPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        parentId:
          title: Parentid
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
      description: Put category request body
    ReqFavorite:
      title: ReqFavorite
      required:
        - user
        - video
      type: object
      properties:
        user:
          title: User
          type: string
        video:
          title: Video
          type: string
      description: Favorite request body
...
    ReqLikeDelete:
      title: ReqLikeDelete
      required:
        - video
        - id
      type: object
      properties:
        video:
          title: Video
          type: string
        id:
          title: Id
          type: string
      description: Delete like request body
    ReqLikePost:
      title: ReqLikePost
      required:
        - video
        - user
        - like
      type: object
      properties:
        video:
          title: Video
          type: string
        user:
          title: User
          type: string
        like:
          title: Like
          type: boolean
      description: Post like request body
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...